vfs fully functional

+++ vfsd/vfs.c
@@ -4,6 +4,7 @@
  */
 
 #include <stdlib.h>
+#include <stdio.h>
 #include <vfsd.h>
 
 void vfs_add(struct vfs *root, const char *path, uint32_t server, uint64_t inode) {
@@ -12,11 +13,11 @@ void vfs_add(struct vfs *root, const char *path, uint32_t server, uint64_t inode
 	}
 
 	if (path[0]) {
-		if (!root->next[path[0]]) {
-			root->next[path[0]] = calloc(sizeof(struct vfs), 1);
+		if (!root->next[(size_t) path[0]]) {
+			root->next[(size_t) path[0]] = calloc(sizeof(struct vfs), 1);
 			root->refc++;
 		}
-		vfs_add(root->next[path[0]], &path[1], server, inode);
+		vfs_add(root->next[(size_t) path[0]], &path[1], server, inode);
 	}
 	else {
 		root->server = server;
@@ -30,8 +31,8 @@ struct vfs *vfs_get(struct vfs *root, const char *path) {
 	}
 
 	if (path[0]) {
-		if (!root->next[path[0]]) return NULL;
-		else return vfs_get(root->next[path[0]], &path[1]);
+		if (!root->next[(size_t) path[0]]) return NULL;
+		else return vfs_get(root->next[(size_t) path[0]], &path[1]);
 	}
 	else {
 		return root;
@@ -44,11 +45,11 @@ void  vfs_lnk(struct vfs *root, const char *path, struct vfs *link) {
 	}
 
 	if (path[0]) {
-		if (!root->next[path[0]]) {
-			root->next[path[0]] = calloc(sizeof(struct vfs), 1);
+		if (!root->next[(size_t) path[0]]) {
+			root->next[(size_t) path[0]] = calloc(sizeof(struct vfs), 1);
 			root->refc++;
 		}
-		vfs_lnk(root->next[path[0]], &path[1], link);
+		vfs_lnk(root->next[(size_t) path[0]], &path[1], link);
 	}
 	else {
 		root->link = link;
@@ -65,12 +66,12 @@ bool vfs_rem(struct vfs *root, const char *path) {
 	}
 
 	if (path[0]) {
-		if (!root->next[path[0]]) return 0;
+		if (!root->next[(size_t) path[0]]) return 0;
 		else {
-			f = vfs_rem(root->next[path[0]], &path[1]);
+			f = vfs_rem(root->next[(size_t) path[0]], &path[1]);
 			if (f) {
-				free(root->next[path[0]]);
-				root->next[path[0]] = NULL;
+				free(root->next[(size_t) path[0]]);
+				root->next[(size_t) path[0]] = NULL;
 				root->refc--;
 			}
 			if (!(root->refc || root->server || root->link)) {
